Split attention.ts and analytics.ts into sub-modules
attention.ts (2,738 lines) → thin orchestrator + 4 sub-modules:
- attention/types.ts: all types, constants, and shared helpers
- attention/project-fields.ts: issue project snapshot resolution
- attention/queries.ts: all DB fetch functions
- attention/mentions.ts: unanswered mention fetch and filtering

analytics.ts (1,954 lines) → existing sub-modules + new transformer:
- analytics/transformers.ts: mapRepoDistribution, mapRepoComparison,
  mapReviewerActivity, mapLeaderboard, mapPrCompletionLeaderboard, toTrend

Both orchestrators re-export everything from sub-modules to preserve
all existing import paths.

Closes #367
